Avoid temporary file for primitive input
This uses primitives new `-o -` option to print the resulting SVG to
stdout. That means we don't need to use a temporary file to capture the
output, which both means one less dependency and running a tad bit
faster. Though to be fair, the run time for primitive dwarfs pretty much
all IO.

https://github.com/fogleman/primitive/pull/71